"Implemented authentication and restructured files"
This commit removes the auth.global.ts middleware and replaces it with a new auth.ts middleware file. The use of useAuth() was also moved from pages to the auth middleware. About page was simplified, removing the need for authentication check inside it. Todo page was changed to fetch data directly from an API instead of using the todo store. The layouts for login and the default were added, and the Nuxt configuration was changed to disable server side rendering. Login page and useAuth composable were heavily revised to properly handle user authentication, user details, and redirection based on authentication status. README was also updated to reflect these changes. This streamlines the authentication process and makes better use of Vue's composition API and Nuxt's middleware, improving code organisation and readability.